make ErrorCode 32bit

+++ Framework/Include/ErrorCode.hpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 #include <sstream>
 
-enum class ErrorCode : uint8_t
+enum class ErrorCode : uint32_t
 {
 	UNKNOWN_ERROR = 0,
